+++ js/app/crew_route/EditCrewRouteCtrl.js
@@ -66,6 +66,12 @@ function EditCrewRouteCtrl(
                     latitude: estimate.job_lat,
                     longitude: estimate.job_lng
                 };
+                estimate.markerEvents = {
+                    click: function() {
+                        $scope.clearHighlights();
+                        estimate.highlight = true;
+                    }
+                };
                 estimate.total = parseFloat(estimate.total);
                 $scope.assignedEstimates.push(estimate);
                 $scope.currentAssignedEstimates.push(estimate);
@@ -111,6 +117,12 @@ function EditCrewRouteCtrl(
                     estimate.coords = {
                         latitude: estimate.job_lat,
                         longitude: estimate.job_lng
+                    };
+                    estimate.markerEvents = {
+                        click: function() {
+                            $scope.clearHighlights();
+                            estimate.highlight = true;
+                        }
                     };
                     estimate.total = parseFloat(estimate.total);
                     $scope.pendingEstimates.push(estimate);
@@ -263,8 +275,23 @@ function EditCrewRouteCtrl(
         $scope.drawRouteDirection();
     };
 
-    $scope.openMarker = function() {
-        // TODO: implement
+    $scope.openMarker = function(estimate) {
+        estimate.show_infor_window = true;
+        $scope.map.control.getGMap().setCenter(
+            new google.maps.LatLng(
+                estimate.coords.latitude,
+                estimate.coords.longitude
+            )
+        );
+    };
+
+    $scope.clearHighlights = function() {
+        for (var i = 0; i < $scope.pendingEstimates.length; i++) {
+            $scope.pendingEstimates[i].highlight = false;
+        }
+        for (var j = 0; j < $scope.assignedEstimates.length; j++) {
+            $scope.assignedEstimates[j].highlight = false;
+        }
     };
 
     $scope.saveRoute = function() {